+++ CHANGELOG.md
@@ -15,7 +15,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Fixed
 
-* Entity usings for data annotations (fixes #126)
+* Entity usings for data annotations (#126)
+* get all is plural (#127)
 
 ## [0.23.0] - 09/22/2023